fix(shell): C3 Codex R3 — f64 dppx precision + refresh-cell-before-spawn root fix
Codex round-3 on PR #415. 2 IMPORTANT, both fixed (one a ≥2-round root fix).

A (dppx f32 precision) — `window.devicePixelRatio` and `@media (resolution)` were
fed an f32 dppx, rounding a fractional scale like 1.2 to 1.2000000476837158, so
JS observed a wrong DPR (CSSOM View says devicePixelRatio is a double). Carry the
scale as lossless f64 end-to-end: ContentAreaPlacement.scale_factor (the D6 single
source) → DeviceFacts.dppx → SetViewport/SetDeviceFacts → the bridge → the
devicePixelRatio getter. The compositor transform / input ÷scale tolerate f32, so
the narrowing is centralized in one `ContentAreaPlacement::scale_f32()` helper
(origin_phys/size_phys/ContentPlacement/scroll); the input mapper (cursor_to_content)
now divides by the f64 directly (a more precise inverse).

  Re-examined the @media half of the finding: with cssparser tokenizing a
  `<dimension>` value as f32 (parse.rs → `Dppx(f64::from(f32))`) and `RangeValue::Dppx`
  comparing EXACTLY, an f64 device 1.2 would never equal the f32-sourced query 1.2 —
  the f32 device coincidentally matched, and a naive f64 device would have *regressed*
  `(resolution: 1.2dppx)`. So the media_environment narrows the device dppx through f32
  to share the cssparser precision (Codex's "normalize before feeding MediaEnvironment"
  alternative) — devicePixelRatio stays exact f64, the media comparison stays
  consistent. Regression test: a delivered 1.2 dppx → devicePixelRatio === 1.2 AND
  (resolution: 1.2dppx) matches.

B (publish before browser-side tab spawns) — ≥2-round root fix. R2's #4a publish-
before-drain covered window.open()-via-drain (drain runs after the redraw-top
publish); Codex R3 found the sibling entry: a browser-side command (Ctrl+T via the
keyboard handler, in the event→redraw gap) calls open_new_tab, which clones the
viewport_cell for the spawned content thread while the resize/DPI arm has deferred
the publish to redraw — so the new tab is born at the stale pre-resize
innerWidth/devicePixelRatio. Root: a construction-input spawn can read a stale cell
because the publish is deferred (D2). Fix: extract the redraw-top recompute+publish
into `refresh_viewport_cell` (idempotent, placement+seq atomic) and call it at BOTH
chokepoints — redraw-top (before the drain) and open_new_tab (before cloning the
cell) — so the cell holds the settled state before any spawn reads it. Subsumes both
entries by construction. (Browser-thread spawn isn't headlessly testable — no winit
window in the content-test harness; structural-by-construction + manual smoke, per
plan-memo §6's stance on window-level behavior.)

+++ crates/script/elidex-js-boa/src/bridge/mod.rs
@@ -113,8 +113,10 @@ pub(crate) struct HostBridgeInner {
     /// Cached viewport dimensions (set by content thread on `SetViewport`).
     viewport_width: f32,
     viewport_height: f32,
-    /// Device pixel ratio (set by content thread from winit `scale_factor`).
-    device_pixel_ratio: f32,
+    /// Device pixel ratio (set by content thread from winit `scale_factor`). `f64`
+    /// (lossless — an `f32` rounds a fractional scale like 1.2, which JS observes as a
+    /// wrong `devicePixelRatio` and the exact `@media (resolution)` evaluator rejects; C3 R3).
+    device_pixel_ratio: f64,
     /// `prefers-color-scheme` user/OS preference (set by content thread from the
     /// window theme, C3). Feeds the [`MediaEnvironment`] the canonical evaluator
     /// reads for `@media (prefers-color-scheme)`; defaults `Light`.
